+++ src/test/scala/scodec/codecs/ChecksumCodecTest.scala
@@ -3,33 +3,35 @@ package scodec.codecs
 import scodec._
 import scodec.bits._
 
-/**
- * ChecksumCodecTest
- */
 class ChecksumCodecTest extends CodecSuite {
 
   "checksummed codec" should {
-    val codec = checksummed(ChecksumCodec.xor(1, int32, 4), variableSizeBytes(int32, utf8))
+    def xor(length: Long) = (bits: BitVector) => bits.grouped(length).foldLeft(BitVector.low(length))(_ xor _)
+
+    val codecSizeIncluded = checksummed(utf8_32, xor(8), peekVariableSizeBytes(int32) ~ bits)
+    val codecSizeExcluded = checksummed(utf8, xor(8), variableSizeBytes(int32, bits) ~ bits)
 
     "roundtrip" in {
-      forAll { (s: String) => roundtrip(codec, s) }
+      forAll { (s: String) => roundtrip(codecSizeIncluded, s) }
+      forAll { (s: String) => roundtrip(codecSizeExcluded, s) }
     }
 
     "roundtrip using combinators" in {
-      forAll { (n: Int, s: String) => roundtrip(int32 ~ codec, n ~ s) }
+      forAll { (n: Int, s: String) => roundtrip(int32 ~ codecSizeIncluded, n ~ s) }
+      forAll { (n: Int, s: String) => roundtrip(int32 ~ codecSizeExcluded, n ~ s) }
     }
 
     "append checksum on encode" in {
-      codec.encode("hello world").require should equal(hex"0x0000000b68656c6c6f20776f726c642b".bits)
+      codecSizeIncluded.encode("hello world").require shouldBe hex"0x0000000b68656c6c6f20776f726c642b".bits
     }
 
     "verify (and remove) checksum on decode" in {
-      codec.decode(hex"0x0000000b68656c6c6f20776f726c642b".bits).require.value should equal("hello world")
-      codec.decode(hex"0x0000000b68656c6c6f20776f726c642b".bits).require.remainder should equal(BitVector.empty)
+      codecSizeIncluded.decode(hex"0x0000000b68656c6c6f20776f726c642b".bits).require.value shouldBe "hello world"
+      codecSizeIncluded.decode(hex"0x0000000b68656c6c6f20776f726c642b".bits).require.remainder shouldBe BitVector.empty
     }
 
     "fail decoding on checksum mismatch" in {
-      codec.decode(hex"0x0000000b68656c6c6f20776f726c6400".bits) should equal(Attempt.failure(ChecksumCodec.Mismatch(hex"0x0000000b68656c6c6f20776f726c64".bits, hex"2b".bits, hex"00".bits)))
+      codecSizeIncluded.decode(hex"0x0000000b68656c6c6f20776f726c6400".bits) shouldBe Attempt.failure(ChecksumMismatch(hex"0x0000000b68656c6c6f20776f726c64".bits, hex"2b".bits, hex"00".bits))
     }
   }
 }